made Clique::getModeLabel() to be const, added Pn potts capabilities
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18308 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 7f65d6a5c19b3b1265464f125e5a8c5a11d78ddf

+++ sandbox/functional_m3n/include/functional_m3n/m3n_params.h
@@ -122,14 +122,14 @@ class M3NParams
      *                            Each value must be in the range [0, 0.5[
      */
     // --------------------------------------------------------------
-    int setInferenceRobustPotts(const vector<double>& robust_potts_params);
+    int setInferenceRobustPotts(const vector<float>& robust_potts_params);
 
     // --------------------------------------------------------------
     /**
      * \brief
      */
     // --------------------------------------------------------------
-    const vector<double>& getRobustPottsParams() const;
+    const vector<float>& getRobustPottsParams() const;
 
     //@}
 
@@ -166,7 +166,7 @@ class M3NParams
     unsigned int nbr_iterations_;
 
     // Inference related
-    vector<double> robust_potts_params_;
+    vector<float> robust_potts_params_;
 
     // Regressor related
     RegressorWrapper::algorithm_t regressor_algorithm_;
